build custom code functions project on startup, update webjobs sdk version, update custom code debugger text
