+++ apps/vs-code-designer/src/app/commands/createNewCodeProject/CodeProjectBase/SetWorkspaceSettings.ts
@@ -76,7 +76,7 @@ export class SetWorkspaceSettings extends AzureWizardPromptStep<IProjectWizardCo
    * @param workspacePath - Root path of the workspace
    */
   private async setupCustomDirectories(context: IProjectWizardContext, workspacePath: string): Promise<void> {
-    const functionFolderPath = path.join(workspacePath, context.methodName);
+    const functionFolderPath = path.join(workspacePath, context.functionAppName);
     await fs.ensureDir(functionFolderPath);
     context.functionFolderPath = functionFolderPath;
   }
@@ -90,7 +90,7 @@ export class SetWorkspaceSettings extends AzureWizardPromptStep<IProjectWizardCo
     const workspaceFolders = [];
 
     // Add Functions folder first if it's a custom code code Logic App
-    const functionsFolder = context.methodName;
+    const functionsFolder = context.functionAppName;
     if (context.isWorkspaceWithFunctions) {
       workspaceFolders.push({ name: functionsFolder, path: `./${functionsFolder}` });
     }
@@ -102,6 +102,12 @@ export class SetWorkspaceSettings extends AzureWizardPromptStep<IProjectWizardCo
     const workspaceData = {
       folders: workspaceFolders,
     };
+
+    if (context.isWorkspaceWithFunctions) {
+      workspaceData['settings'] = {
+        'azureFunctions.functionFolderPath': context.functionFolderPath,
+      };
+    }
 
     await fs.writeJSON(context.workspaceCustomFilePath, workspaceData, { spaces: 2 });
   }
@@ -118,7 +124,7 @@ export class SetWorkspaceSettings extends AzureWizardPromptStep<IProjectWizardCo
     const workspaceFolders = [];
 
     // Add Functions folder first if it's a custom code code Logic App
-    const functionsFolder = context.methodName;
+    const functionsFolder = context.functionAppName;
     if (context.isWorkspaceWithFunctions) {
       workspaceFolders.push({ name: functionsFolder, path: `./${functionsFolder}` });
     }
@@ -128,6 +134,14 @@ export class SetWorkspaceSettings extends AzureWizardPromptStep<IProjectWizardCo
     workspaceFolders.push({ name: logicAppName, path: `./${logicAppName}` });
 
     workspaceContent.folders = [...workspaceContent.folders, ...workspaceFolders];
+
+    // TODO: Support for multiple custom code functions projects
+    if (context.isWorkspaceWithFunctions) {
+      workspaceContent.settings = {
+        ...workspaceContent.settings,
+        'azureFunctions.functionFolderPath': context.functionFolderPath,
+      };
+    }
 
     await fs.writeJSON(context.workspaceCustomFilePath, workspaceContent, { spaces: 2 });
   }
