+++ apps/vs-code-designer/src/app/commands/createNewCodeProject/createCodeProjectSteps/createFunction/FunctionAppNamespaceStep.ts
@@ -6,7 +6,7 @@ import { localize } from '../../../../../localize';
 import { AzureWizardPromptStep } from '@microsoft/vscode-azext-utils';
 import type { IProjectWizardContext } from '@microsoft/vscode-extension-logic-apps';
 
-export class setNamespace extends AzureWizardPromptStep<IProjectWizardContext> {
+export class FunctionAppNamespaceStep extends AzureWizardPromptStep<IProjectWizardContext> {
   public hideStepCount = true;
 
   public async prompt(context: IProjectWizardContext): Promise<void> {
